move `get_fints_transactions` to `lib`
This is now possible because the `fints` fetching logic is not under
`web`, but rather `lib.external_services`.
